Return an empty hash from `#flow_session` if it has not been created
We have observed a number of 500 errors that are the result of the flow session being nil. There is a new one present in the document capture controller.

This commit fixes all of these issues and hopefully prevents new ones by modifying the non-FSM implementaitons of `#flow_session` to return an empty hash when the flow session has not been constructed yet.

[skip changelog]