LG-9161 Remove drivers license verification from instant verify
Drivers licenses were previously sent to LexisNexis to compare results to AAMVA to explore the possibility of useing LexisNexis as a fallback from DL verification. We determined that this is not a feasible fallback.

This change removes sending drivers licenses info to LexisNexis to discontinue drivers licenses validation since we will not be using these results.

[skip changelog]