Write to the ThreatMetrix proofing component even if ThreatMetrix is disabled
We use the ThreatMetrix proofing component to determine if additional review is needed for a user. This change ensures that the ThreatMetrix proofing component reflects the status of the last proofing transaction.

E.g. if Threatmetrix is enabled and a user proofs and requires fraud review they will not be required to do fraud review if a later session where ThreatMetrix is disabled passes but the component is not updated.

[skip changelog]