LG-9209 Use the ExecutedStepName to group LexisNexis errors
Previously we used the `ProductType` key in the LexisNexis verification error parser as the key for errors. This was done to support the get-to-yes feature with AAMVA. #8176 changes the get-to-yes feature to inspect the response body instead of the errors hash so this is not longer needed to support get-to-yes.

The `ProductType` key is problematic because the LexisNexis response can have multiple products with the same product type. We have observed this with PhoneFinder where responses may contain 2 PhoneFinder products with different `ExecutedStepName` values. Grouping by `ProductType` causes one of the products to be overwritten. This commit attempts to address that issue by grouping by `ExecutedStepName`.

[skip changelog]