Use the instant verify product instead of the parsed errors for get-to-yes (#8176)
The get-to-yes with AAMVA feature was developed to allow users whose data fails a LexisNexis check to pass if the same attributes pass an AAMVA check. This was done by observing the errors from the LexisNexis transaction and determining which attributes fail and whether they can be covered by AAMVA.

This approaches to determining which attributes fail from the errors turned out to be problematic. The errors are parsed from the LexisNexis InstantVerify product portion of the RDP response and returned as a hash. The structure of that hash has changed occasionally causing this feature to quietly break.

This commit attempts to clean this up by directly inspecting the InstantVerify product instead of looking at the errors hash. This requires parsing out the product in the proofer. There is also some minor refactor to the response object to make some of the attributes in the response first class methods there.

[skip changelog]