LG-9065 Include TMx response in the resoltuion result adjudicator logic (#8163)
The `ResolutionResultAdjudicator` class is responsible for inspecting the results we get from making proofing API calls. It makes an adjudication on the proofing transaction based on those results.

In the past the ThreatMetrix result has not been included in this adjudication. This was because the transaction was consider successful regardless of what happened with ThreatMetrix. This was possible because ThreatMetrix results were handled downstream later in the proofing process.

Now, we want to consider cases where ThreatMetrix encounters an exception in the adjudication. This is because we want transactions where ThreatMetrix encounters an exception to be considered failures so the user will retry instead of failing downstream due to something like a network timeout or 500 from the TMx API.

[skip changelog]