LG-9065 Add the ThreatMetrix review status to the result of the resolution proofing job (#8164)
We have been writing the ThreatMetrix review status to the proofing components in the database. This has some issues because we need to make sure that the results are not overwritten before we read them to make a determination. This is especially problematic for the GPO letter flow where the results are inspected days after they are written.

This commit adds the review status to the result of the resolution proofing job. This will allow the proofing flow to hold it in the session and use it to set a value in the `fraud_review_pending` column when creating a profile.

[skip changelog]