Copy override_document_capture_step_csp before action into DocumentCaptureController
This method allows Acuant to work by adding a new ContentSecurityPolicy. As part of the Flow State Machine it lives
in DocumentCaptureConcern, and checks the step param for document_capture. Since the new controller is only for the
DocumentCapture step, we can add the method directly into the controller. We can't use the one in DocumentCaptureConcern
because the new controller doesn't have a step param.