Remove VerifyInfoController step counts
The Flow State Machine kept count of the number of times a screen was visited in the course
of one session. We are not using that information in any dashboards, so remove it from the
post-FSM controllers.

changelog: Logging, Identity Verification, remove step counts from analytics for steps outside the Flow State Machine