LG-8617 Add a boolean column to mark that a profile is pending GPO review
As part of our profile state work we are unloading the `deactivation_reason` column. This involves adding new columns to represent whether a profile satisfies one of several conditions that renders it inactive. This is necessary since as time has gone on we have a number of temporary conditions that render a profile inactive and can exist simultaneously.

This commit adds a column for tracking that a user is pending GPO review. We cannot start writing to this column until after this migration has been deployed and run in production.

[skip changelog]