Set :same_address_as_id = false on update res address
Ideally we'd fix this field to be false on the form submission, but the
form doesn't have an `updating_address?` check. The easiest way to
implement a fix is in `address_step#call`.

We were previously skipping setting the `same_address_as_id` field
because if it isn't skipped, `same_address_as_id` gets set to `nil`.

The `if` statement is a little verbose, but after talking with Tomas,
we decided the verbosity is worth the clarity it brings. Also, once the
feature is enabled fully, we can remove one of the three checks in the
`if` statement.