LG-9726 Introduce a fraud review check class
We have a bunch of disparate logic around checking whether a user has or should be deactivate for fraud.

This commit introduces the `FraudReviewCheck` class. The intention of this class is to provide a place to consolidate logic to determine if a user has a pending fraud review check.

I am planning to follow this up with logic that determines if a user should receive a timestamp update for a pending fraud review check.

[skip changelog]