Change the fraud review logic for fraud review
We have been making changes to the way that we manage users in fraud review. Specifically we have been unloading the `deactivation_reason` column and using specific columns each reason a profile could be in review.

This proves problematic in some cases since some of the logic that assumes a single deactivation reason is still in place. Specifically, a user can now be in fraud review or GPO pending state simultaneously, but the logic for changing states does not take that into consideration. As a result the logic for moving users in between states is incomplete. This commit helps to alleviate that by adjust the fraud review profile query to only return a fraud review profile when a user is not in a GPO pending state and their profile has not been deactivate for some other reason.

[skip changelog]