Check for a deactivation reason before activating a profile
We should not be activating profiles that have a deactivation reason. The logic for that looks like it got mixed up when we were doing the fraud review work. Specifically the code to check for a deactivation reason does not actually look at the deactivation reason column anymore.

This commit addresses the above.

[skip changelog]