Check for a deactivation reason before activating a profile (#8426)
We should not be activating profiles that have a deactivation reason. The logic for that looks like it got mixed up when we were doing the fraud review work. Specifically the code to check for a deactivation reason does not actually look at the deactivation reason column anymore.

This commit addresses the above.

[skip changelog]

(cherry picked from commit df5a9054cd71370cc2243a577f2e723bf647dc04)