LG-9766: Update verify page headings for accessibility (#8415)
changelog: User-Facing Improvements, In-person proofing, Update verify page headings for accessibility

Currently, when using a screen reader, the only heading is the `h1`. We
wanted the `state-issued ID`, `current residential address`, and
`social security number` headings to also be read as headings, but not
change the styling.

The styling was very similar to an h4. We couldn't use an `h4` element
because while it would add landmarks to the page, it would make screen
reader users potentially concerned they are missing content to go from
an `h1` to an `h4`, with no `h2` or `h3` between.