Do not return cancelled or errored profiles from `#pending_profile`
The pending profile will return a profile that is pending GPO confirmation, pending in-person, or pending fraud review.

In the above cases, if a user cancels verification or has an encryption error then their profile is no longer verifiable and is not pending. This commit adds a guard to the `pending_profile` method to return nil in these cases.

[skip changelog]