Make the confirrm start over change backwards compatible
Previous commits introduced a new screen for confirming you want to cancel IdV. Unfortunately this is not backwards compatible with the 50/50 state. The proper way to deploy this is:

1. Add the new page at an accessible route
2. Start redirecting users to the new route
3. Tear down any code that supported the old approach

This commit winds everything back to step 1 by making the new functionality available but not redirecting users to the new page.

Note that this particular change does not need step 3 since there actually is not an old route to remove. This adds an interstitial page that contains a form to make a request to a route that users are already using.

[skip changelog]