Brewfile for dependency management
*Summary of changes*

We're looking at adding a dependency soon, which could cause some issues
with local development for folks if they didn't know they need to
manually install it. We thought adding a `Brewfile` and having the setup
script run `brew bundle` would help mitigate the difficulties of that
change. There's a bonus in that it's self documenting too!

It would also:

- make initial setup for new developers much more straightforward
- make future changes to dependencies easier to manage
- prevents a mixture of dependencies installed in different ways (for
instance, I think when I initially set things up I ended up with one
version of postgres downloaded manually from the doc-linked site, and
another via homebrew)

Notes:

- As pointed out by @pauldoomgov, we use [postgres@13 in production](https://github.com/18F/identity-devops/blob/33ba02736b37f3a79c50479892a03c6f3e920041/terraform/app/rds-variables.tf#L82),
so it would be good to encourage use of 13 locally. The method I chose
to try and enforce this was using `brew services stop --all` before
starting the services we want (where I specify 13).
- I tried various combinations of having these dependencies installed /
uninstalled before running `make setup` and the script seems to still
work. But I don't have a machine without MacOS to test on to make sure
it works for everyone.
- While making relevant doc changes, I also fixed up some syntax for
readability / easy re-ordering of steps