Change `pending_profile_requires_verification?` to only apply to GPO verification
We re-implemented pending_profile to return profiles that are pending for the following reasons:

1. In-person proofing
2. GPO verification
3. Fraud review

The `pending_profile` method at one point only returned GPO pending profiles and thus `gpo_verification_pending_profile?` is used in a bunch of contexts that apply specifically to GPO. This commit narrows the scope for this method.

changelog: Improvements, Profile state, The `pending_profile_requires_verification?` method was changed to describe any pending state but was still used in contexts that redirect for GPO. This method was changed to only apply to GPO pending users.