Use the presence of the flow path in the session as a proxy for upload step completion (#8456)
* Use the present of the flow path in the sessino as a proxy for the upload step complete

The document capture controller no longer uses the FSM. It does, however, check that the upload step in the flow state machine is complete. Prior to this commit it checks that the step is marked complete in the flow session. This means that we need to continue marking the step complete in an eventual upload step replacement.

This commit changes the way that we check that the step is complete to look that the `flow_path` has been set. This brings the upload step completion in line with the new link sent controller.

[skip changelog]