Sign in live in the in person happy path tests
The in-person tests were using the `sign_in_user` helper to test that a user sees the screen with their barcode visible. This signs in with Warden which does not run through all of the before actions for signing in.

Additionally, when the in person happy path tests were signing in the user was already signed in.

This commit addresses both of the above.

[skip changelog]