Use the verify profile concern to sort through all of the reasons a user may be pending (#8468)
We have made changes to the `pending_profile` concept in the IdP. Previously a profile was pending if the user was waiting on GPO verification. We have added more out of band flows which result in more pending reasons:

1. GPO Verification
2. Fraud review
3. In-person proofing

This logic got spread throughout the codebase in ways that were often kind of tough to follow. This led to lots of bugs trying to send the user to the right place to finish proofing.

This commit collects that logic in the `VerifyProfileConcern`. This was chosen because this is where much of the logic for the GPO confirmation flow lived already.

[skip changelog]