Allow using StepIndicatorConcern w/o IdvSession
If the user is not logged in and does not have an `idv_session` available (as is the case in the hybrid document capture flow), this code would raise and the user would get a 500 error.

This update makes an assumption that if the user is not properly logged in they are probably not in the gpo or in person flow.

[skip changelog]