Removing acuant a/b test reporting from hh analytics
-- Why
The original upload step spec did not test for the presence of the
acuant analytics at all. We added this check to the _new_ controller
test, but it's unclear that we want to even log acuant ab testing
information at this step.

We are also restoring the removed code in the acuant concern which
checked for the presence of a document capture session uuid before
returning any value