LG-9959 Fix a bug in which in-person and GPO pending users do not have GPO pending status cleared (#8492)
We had a bug where users who are in the GPO pending state did not have that state cleared after entering a successful code and going through in-person proofing. This resulted in an error attempting to activate these users profiles when the enrollments background job runs.

changelog: Improvements, In-person proofing, A bug in which users who were letter and in-person pending did not have their letter pending status cleared after entering a letter code resulting in the user not being proofed after visiting the post office was fixed.