Log redo_document_capture in document_capture_visited
We want to log when redo_document_capture happens because it sometimes causes unexpected results
for the user. When it was its own action it had its own analytics event, but now we are combining
it with document_capture_visited. It will also be added to hybrid_handoff for users who are not
on mobile.

changelog: Internal, Identity verification logging, log redo_document_capture from document_capture and hybrid_handoff

Co-authored-by: Douglas Price <douglas.price@gsa.gov>