Add tests to cover password resets with a pending profile (#8628)
We have uncovered a number of bugs that occur when a user with a pending profile resets their password. This commit adds tests to cover this behavior and make sure everything works as expected.

Currently if a user with a pending profile resets their profile we deactivate their profile and it is unrecoverable. These tests may need to change if we give users the option to recover these profiles.