LG-9727 Add a fraud_pending_reason_column
This commit adds a column for holding the reason that a user is fraud pending. This commit also clears the fraud review pending reason if it exists when the profile is activated after fraud review.

In order to maintain consistency the commit that adds the fraud review pending reason will need to follow this one in a different deploy.

changelog: Internal, Fraud Review, A fraud_pending_reason column was added to track the reason that a user entered the fraud review process.