Install audit-ci as devDependency
It's not the "recommended" usage per docs, but our CI setup is such that the lint would run after the initial installation anyways (per the documented "downsides"). This way, we also avoid potential issues with vulnerable dependencies of audit-ci itself, since while we can pin audit-ci version, audit-ci's dependencies use loose version ranges. With lockfile, we also expect that dependencies shouldn't change from under us. What we're missing is the ability to know that a newly-introduced package would be vulnerable prior to running the installation.

https://github.com/IBM/audit-ci#set-up