LG-9818 Set the `address_verification_mechanism` for the GPO verify action
The personal key controller was previously always part of the main proofing flow and was always rendered after the password input. Recently we moved it to after GPO verification. This meant that when the user was seeing their personal key for the GPO verification workflow the Idv::Session had been cleared when their session expried. As a result the GPO verification screen needs to setup an Idv::Session for the personal key.

When this was done the `address_verification_method` attribute was left off. The personal key controller users the `StepIndicatorConcern` to determin which steps to show. That looks for the presence of a pending profile or that attribute on `idv_session`. Since the user no longer has a pending profile after confirming their GPO code and the GPO verification controller was not setting `address_verification_mechanism` the `StepIndicator` concern was using the default steps instead of the GPO specific ones.

This commit fixes the issue by setting `address_verification_method` when the user successfully confirm their code.

changelog: Improvements, Verify by mail, A bug in which the user was displayed the default step indicator steps instead of the verify-by-mail step indicator steps in the step indicator on the personal key issuance screen after confirming the OTP as part of the verify by mail flow was fixed.