LG-9969 Update the `fraud_pending_reason` backfill rake task
This commit updates the rake task that updates `fraud_pending_reason`. It makes the following changes:

1. The `backfill_fraud_pending_reason` task performs a dry run by default and needs a flag set to commit changes. This allows for a dry-run to be done to preview changes that will be made.
2. A task is added to rollback the operation in case that is necessary.

changelog: Internal, Fraud review, The rake task that updates the fraud_pending_reason column was changed such the the taks for backfilling the column has additional output and does a dry run by default and such that a rollback task was added to perform the reverse of the backfill operation given the backfill operations output.