LG-9806 Do not redirect to address controller when a PR user updates their address
Users who upload a Puerto Rican document during IdV are redirected to the address controller after they enter their SSN. This differs from the rest of users who are sent to the verify info screen and only see the address controller views if they choose to update their address there. This is because Puerto Rican addresses have been problematic and seeing the address screen gives users with Puerto Rican documents a change to modify their address if it is read incorrectly. There are also alerts specific to formatting PR addresses that appear here.

To enable the above behavior the SSN controller redirects to the address controller if the jurisdiction for the user's address is "PR". This approach worked well for PR users who are entering their SSN for the first time. This was problematic for Puerto Rican users who are updating their SSN after reaching the verify info screen. They would be sent to the address screen after updating their SSN even though they had already seen the address screen on their way to the verify info screen originally.

This commit address the problematic behavior by also checking whether the user is updating their SSN before the redirect to the address controller. Changes were necessary in `SsnFormatForm` to allow this. Previously the `updating_ssn?` method looked for the presence of an SSN to determing if the SSN was being updated. This does not work in the PR case since after a valid submission the form always has an SSN. Instead `updated_ssn?`'s return value is set in the initializer when the existing SSN is passed in as a parameter if it exists.

changelog: Improvements, Puerto Rican Proofing Experience, Users who are proofing and upload a Puerto Rican document during document capture and do not change their jursidiction on the address update page are no longer redirected to the address update page after submitting an ssn after choosing to update their ssn from the verify info page.