Enable the v3 session encryptor by default
We added the v3 session encryptor a while back. This conifg was used to migrate away from the v2 encryptor to the v3 encryptor. This commit flips it on to default not that that migration is done.

This change is in anticipation of removing the legacy session encryptor code which can safely be done 1 release cycle after this change.

changelog: Internal, Session encryption, The legacy session encryptor which was previously enabled by default was disabled in favor of the v3 session encryptor.