Retire old MultiRegionKmsClient
We implemented a version of a multi-region KMS client. It was intended to encrypt across multiple regions so that if a region was unavailable a ciphertext could be decrypted in a different region. This code was never enabled due to complications with configuring the KMS clients.

Since this implementation was put together AWS has released KMS instances that are multi-region. This means we no longer need to use this approach to encrypt/decrypt across regions. As a result, this code can be removed.

[skip changelog]